refactor: introduce 도메인 구조 정리 (sns 패키지 이동, intro 제거)
