--- src/main/java/com/example/cowmjucraft/domain/introduce/intro/entity/IntroduceContent.java
@@ -1,43 +1,0 @@
-package com.example.cowmjucraft.domain.introduce.intro.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Lob;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "introduce_contents")
-public class IntroduceContent {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Lob
-    @Column(nullable = false)
-    private String content;
-
-    @Column
-    private Long logoMediaId;
-
-    @Column
-    private Long bannerMediaId;
-
-    public IntroduceContent(String title, String content, Long logoMediaId, Long bannerMediaId) {
-        this.title = title;
-        this.content = content;
-        this.logoMediaId = logoMediaId;
-        this.bannerMediaId = bannerMediaId;
-    }
-}
