+++ src/main/java/com/example/cowmjucraft/domain/sns/controller/client/SnsController.java
@@ -1,7 +1,7 @@
-package com.example.cowmjucraft.domain.introduce.sns.controller.client;
+package com.example.cowmjucraft.domain.sns.controller.client;
 
-import com.example.cowmjucraft.domain.introduce.sns.dto.response.SnsResponseDto;
-import com.example.cowmjucraft.domain.introduce.sns.service.SnsService;
+import com.example.cowmjucraft.domain.sns.dto.response.SnsResponseDto;
+import com.example.cowmjucraft.domain.sns.service.SnsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
