+++ src/main/java/com/example/cowmjucraft/domain/sns/dto/request/SnsAdminRequestDto.java
@@ -1,6 +1,6 @@
-package com.example.cowmjucraft.domain.introduce.sns.dto.request;
+package com.example.cowmjucraft.domain.sns.dto.request;
 
-import com.example.cowmjucraft.domain.introduce.sns.entity.SnsType;
+import com.example.cowmjucraft.domain.sns.entity.SnsType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
