+++ src/main/java/com/example/cowmjucraft/domain/sns/repository/SnsLinkRepository.java
@@ -1,6 +1,6 @@
-package com.example.cowmjucraft.domain.introduce.sns.repository;
+package com.example.cowmjucraft.domain.sns.repository;
 
-import com.example.cowmjucraft.domain.introduce.sns.entity.SnsLink;
+import com.example.cowmjucraft.domain.sns.entity.SnsLink;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
